pyproject.toml
[tool.poetry]
name = "adrift"
version = "0.1.0"
description = ""
authors = ["Philip De Lorenzo <philip.delorenzo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.41.1"
st-pages = "^1.0.1"
extra-streamlit-components = "^0.1.71"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
watchdog = "^6.0.0"
streamlit-dynamic-tabs = "^0.0.3"
streamlit-dynamic-filters = "^0.1.9"
streamlit-multipage = "^0.0.18"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
